+++ src/main/env.d.ts
@@ -15,6 +15,7 @@ interface ImportMetaEnv {
   readonly VITE_API_ENDPOINT_RECORD_DETAIL: string
   readonly VITE_API_ENDPOINT_RECORD_STATS: string
   readonly VITE_API_ENDPOINT_RECORD_STATS_NEW: string
+  readonly VITE_API_ENDPOINT_RECORD_STATS_BATCH: string
   readonly VITE_API_ENDPOINT_RECORD_EVALUATE: string
   readonly VITE_API_ENDPOINT_RECORD_CLAN_INFO: string
   readonly VITE_API_ENDPOINT_RECORD_BATTLE_ENRICH: string